+++ src/Simpletools/Cloud/Google/Storage/Bucket.php
@@ -39,4 +39,57 @@ class Bucket
     {
         self::$_bucket[$this->_bucketName]->exists();
     }
+
+  public function listFiles($directory = '', $sortBy='',$sortDirection ='asc')
+  {
+    // List objects in a specific directory
+    $options = [];
+    if ($directory) {
+      $options['prefix'] = $directory;
+    }
+
+    $objects = [];
+    foreach (self::$_bucket[$this->_bucketName]->objects($options) as $object) {
+      $info = $object->info();
+      $objects[] = [
+        'name' => $object->name(),
+        'created' => strtotime($info['timeCreated'])
+      ];
+    }
+
+    if($objects)
+    {
+      // Sort by creation time
+      if($sortBy == 'created')
+      {
+        usort($objects, function ($a, $b) use ($sortDirection) {
+          if($sortDirection == 'desc')
+            return $b['created'] <=> $a['created'];
+          else
+            return $a['created'] <=> $b['created'];
+        });
+      }
+      elseif($sortBy == 'name')
+      {
+        usort($objects, function ($a, $b) use ($sortDirection) {
+          if($sortDirection == 'desc')
+            return $b['name'] <=> $a['name'];
+          else
+            return $a['name'] <=> $b['name'];
+        });
+      }
+    }
+
+    return $objects;
+  }
+
+  public function getIterator($directory = '')
+  {
+    // List objects in a specific directory
+    $options = [];
+    if ($directory) {
+      $options['prefix'] = $directory;
+    }
+    return self::$_bucket[$this->_bucketName]->objects($options);
+  }
 }